+++ app/src/main/java/ru/bagrusss/revolutdemo/repository/RatesRepository.kt
@@ -2,12 +2,13 @@ package ru.bagrusss.revolutdemo.repository
 
 import io.reactivex.Observable
 import io.reactivex.Single
+import java.math.BigDecimal
 
 /**
  * Created by bagrusss on 13.08.2019
  */
 interface RatesRepository {
-    var currentBaseRate: Pair<String, Double>
-    val currentCostChanges: Observable<List<Pair<String, Double>>>
-    val actualRates: Single<List<Pair<String, Double>>>
+    var currentBaseRate: Pair<String, BigDecimal>
+    val currentCostChanges: Observable<List<Pair<String, BigDecimal>>>
+    val actualRates: Single<List<Pair<String, BigDecimal>>>
 }
